adds documentation to README
NOTES
This is what we have done so far up to Chapter 6 of the tutorial
and we wanted to document it for anyone trying to do the same.

It has a lot of info but not everything for someone new to GNU/Linux
but it can be of some help.